Select kernel version from CLI
A new command line option has been added to select between the global and shared version. The run script now runs both versions and the comparison scripts.

./master_thesis $1 $2 $3 $4 $5 0 ../output
The two versions do not seem to produce quite the same result. The configuation that resulted in the comparison figure shows 34 cells where the maximum error was above the 0.1% tolerance, although just barely. The maximum absolute error was 0.0010000000000331966.