Implement missing kernel
The kernel to initialize kernel data was not implemented and did nothing. As a consequence part of the shared version did not run to utilize the shared memory, although it produced comparable results.

Also increased the number of decimals used to store the density to rule that out as a cause for the inconsistensies between the versions.