Move charge density out of main file
The charge density kernels have now been moved into separate files. The shared version required a class to handle the lifetime of the temporary storage used by radix sort.

Also moved the debug definition to the CMake file and made the run script abort on errors so that, for example, the figure doesn't show when there is a runtime error and no new data was produced.