Initialize particles indices only once
It's not needed to update the initial particle indices more than once since they are not modified by the radix sort.

Also added command line option to save to disk so that you can disable it and save some time, as well as some other minor changes.

Finally removed the python script from the bash run file to save time when running.