Create script to automate validation
I've created a directory for the first validation. In it is a savefile with particle positions, where the filename contains the grid dimensions. The only missing data is the cell size that is implied to be the same as the hardcoded value used in the plotting scripts. The directory also contains a bash script to build the program, run both versions and finally plot and save the charge density figures. Everything in the validation directory is synced with Git, including the output directory normally ignored.

Modified the density plotting script to allow saving the plots and choosing where to look for density data.