Add comparison script to validation 1
The generated data and plots have been moved to a key directory that will represent the true values and results from validation 1. Running the compare script will compare the output of the validation run with the correct densities and print min, max and mean error. If there were any errors a heatmap showing them will be shown.