Add validation 2 scripts
The second validation computes the total particle charge by accumulating the densities and comparing to the product of particle charge and particel count. The difference of these two values is printed to the console.

Also forgot to use the compare script in the validation 1 bash script. But most importantly, fixed an error in one of the coordinates of the "correct" data for validation 1.