Add cnpy library
The cnpy library massively speeds up saving and loading data. npy is a binary format compared to the csv files used previously.

Also, replaced loop summation with numpy.sum, which drastically changed the result. I trust numpy's summation over a manual one so I'll stick to the built-in methods in the future. And I added more prints to the validation 2 Python scripts.